% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadout.R
\name{spreadout}
\alias{spreadout}
\title{spreadout}
\usage{
spreadout(x, case)
}
\arguments{
\item{x}{data set}

\item{case}{setup info}
}
\value{
A numeric vector of observations without ties.
}
\description{
This function unbins data. If qnull is given it uses quantiles, otherwise uniform
}
\examples{
case <- list(B=1000, param = NULL, n = 1000, pnull = function(x, param) punif(x), 
    rnull = function(n, param) runif(n), qnull = function(x, param) qunif(x), 
    est.mle = function(x) NA, nbins = 10)
y=runif(1000)
bins=seq(0, 1, length=11)
counts=hist(y, bins, plot=FALSE)$counts
x=list(bins=bins,counts=counts)
spreadout(x, case)
}
