% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{TS}
\alias{TS}
\title{TS}
\usage{
TS(x, case)
}
\arguments{
\item{x}{data}

\item{case}{setup info}
}
\value{
A numeric vector with the values of various test statistics.
}
\description{
This function finds various gof statistics
}
\examples{
case <- list(B=1000, param = NULL, n = 1000, pnull = function(x, param) 
         punif(x), rnull = function(n, param) runif(n), qnull = function(x, param) 
        qunif(x), est.mle = function(x) NA, nbins = 10)
case$methods=c("KS", "AD", "CdM", "W", "ZA", "ZK", "ZC")
x <- runif(1000)
TS(x, case)
}
